第三章：langchain实战/test-to-sql-main/pyproject.toml
[project]
name = "text2sql-agent"
version = "0.1.0"
description = "A simple text-to-SQL agent using LangChain's create_agent and GLM-5"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.3",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=1.0.6",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.urls]
Repository = "https://github.com/kevinbfrank/text-to-sql-agent"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
